refactor: Create 'GameEnums' to separate 'PlayerType' from 'Constants'
