+++ Includes/Rosetta/RosettaStone.hpp
@@ -70,6 +70,7 @@
 #include <Rosetta/Enums/ActionEnums.hpp>
 #include <Rosetta/Enums/AuraEnums.hpp>
 #include <Rosetta/Enums/CardEnums.hpp>
+#include <Rosetta/Enums/GameEnums.hpp>
 #include <Rosetta/Enums/TargetingEnums.hpp>
 #include <Rosetta/Enums/TaskEnums.hpp>
 #include <Rosetta/Enums/TriggerEnums.hpp>
